requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy<=1.4.44
passlib<=1.7.4
PyJWT<=2.6.0
fastapi-mail<=1.2.0
psycopg2-binary==2.9.3
python-multipart==0.0.5